Fix failing unit test

+++ core/src/main/java/com/lottie4j/core/model/animation/Animated.java
@@ -47,7 +47,7 @@ import java.util.List;
 @JsonInclude(Include.NON_NULL)
 public record Animated(
         @JsonProperty("a")
-        Double animated,
+        Integer animated,
 
         @JsonProperty("k")
         @JsonSerialize(using = KeyframeSerializer.class)
